requirements.txt
asgiref==3.5.2
Django==4.0.5
sqlparse==0.4.2
# django-froala-editor
psycopg2-binary==2.9.3
django-tinymce
pillow